Exclude the generator from Go without context

+++ test.sh
@@ -5,6 +5,9 @@ echo "" > coverage.txt
 
 for d in $(go list ./... | grep -v vendor); do
     go test -v -race -coverprofile=profile.out -covermode=atomic $d
+    if [[ "$NOCONTEXT" == "1" ]] && [[ d =~ "generator" ]]; then
+        continue
+    fi
     if [ -f profile.out ]; then
         cat profile.out >> coverage.txt
         rm profile.out
